+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+- Added support for Bandai FCG (mapper 16)
+
 ### Changed
 - The emulator display can now fill the entire screen
 - Improved compatibility with Android and macOS